Refactor application with new SMS and Mailing modules
The commit removes the AppModule and replaces it with SMS and Mailing modules. This includes all corresponding controllers, services, DTOs and tests for these modules. Also removed unneeded dependencies and installed the "mapped-types" dependency. This move is meant to better structure the applications and provide more specific services for SMS and Mailing operations.